+++ server/api/user/lists/privacy/[id].ts
@@ -1,34 +1,10 @@
 import type { H3Event } from 'h3';
-import { prisma } from '~/utils/server/prisma';
-import { isNext } from '~/utils/server/debug';
+import { getAllPrivateUsers } from '~/utils/server/user';
 
 export default defineEventHandler(async (event: H3Event) => {
     const id = getRouterParam(event, 'id');
 
-    if (isNext()) {
-        const request = await $fetch<{ isPrivate: boolean }>(`https://vatsim-radar.com/api/user/lists/privacy/${ id }`).catch(() => {});
-        if (request?.isPrivate) {
-            return {
-                isPrivate: true,
-            };
-        }
-    }
-
-    const dbUser = (await prisma.vatsimUser.findFirst({
-        where: {
-            id,
-        },
-        select: {
-            user: {
-                select: {
-                    id: true,
-                    privateMode: true,
-                },
-            },
-        },
-    }));
-
     return {
-        isPrivate: !!dbUser?.user.privateMode,
+        isPrivate: !!(await getAllPrivateUsers())[id ?? ''],
     };
 });